// assets/Script/GameScene/Enum/Enum.ts
/** 背景音乐路径 */
export enum MusicResUrl {
    Mars = "music/Rise of spirit",
    gameStart = "music/gameStart",
	boom = "music/boom",
	biu = "music/biu"
}

/** 武器类型 */
export enum WeaponType {
	A = 0
}

/** 飞机类型 */
export enum EnemyType {
	small = 0,
	meduim = 1,
	big = 2,
	boss = 3
}

/** 爆炸类型 */
export enum ExplosionType {
	explosion_01 = 0,
	explosion_02 = 1,
	explosion_03 = 2
}

/** 玩家动画 */
export enum playerAni {
	idle = "idle",
	toLeft = "toLeft",
	left = "left",
	toRight = "toRight",
	right = "right",
	hit = "hit"
}

/** 道具类型 */
export enum PropType {
	hp = 0,
	bullet = 1,
	gold = 2,
	nuclearBomb = 3
}